requirements.txt
# Core scraping and web automation
playwright==1.40.0
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
selenium==4.15.0

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Caching and background tasks
redis==5.0.1
celery==5.3.4

# Data processing and matching
pandas==2.1.4
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
loguru==0.7.2
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.13.0
